:bug:[배포]: 올바른 파일명 변경
