+++ src/app/constants/common.ts
@@ -1,0 +1,22 @@
+  export const HEADER_TEXT = [
+    '상품목록',
+    '상품주문',
+    '견적서 조회',
+    '주문내역',
+    '회원가입',
+    '로그인',
+  ];
+
+  export const FOOTER_TEXT = [
+    '이용약관',
+    '개인정보처리방침',
+    '책임의 한계와 법적고지 ',
+    '고객센터',
+    'JMF',
+    'Copyright',
+    '©JMF Corp.',
+    'All Rights Reserved',
+    'Contributors',
+  ];
+
+  export const CONTRIBUTORS = ['YPYP', 'KimGorok', 'Minkyu'];
